+++ src/utils/getProjectRoot.ts
